finance_control_app_complete/requirements.txt
Flask==2.2.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.46
mysqlclient==2.1.1
itsdangerous==2.1.2
APScheduler==3.10.1
Werkzeug==2.2.3
xhtml2pdf==0.2.11
python-dotenv==1.0.0
email-validator==2.0.0.post2
WTForms==3.0.1
typing-extensions==3.7.4.3
PyInstaller==6.3.0
